Update README.md
This repository contains the implementation, execution, and deployment of a Bi-Disease Prediction Model in Python and by using Machine Learning. The model is drafted to predict the possibility of whether an individual is suffering from Diabetes or Heart Disease based on key health indicators and features or not. The project focuses on harnessing data-driven insights to assist healthcare professionals, individuals, and organizations in the early detection, management, and diagnosis of such chronic diseases.